Models updated and migration added

+++ Maintain.NET/Maintain.NET/Models/MaintenanceTask.cs
@@ -21,7 +21,7 @@ namespace Maintain.NET.Models
         /// Recommended maintenance interval of MaintenanceTask
         /// </summary>
         [Display(Name = "Recommended Interval")]
-        public int RecommendedInterval { get; set; }
+        public long RecommendedInterval { get; set; }
 
         // Navigation Properties
         /// <summary>
